Brute force node version
Rather than load a single .node addon for the exact version of node, the
code will attempt to load all .nodes for the current processor
architecture. This allows you to only include a build for incompatible
versions.

The 0.12 versions have been removed from build-targets.txt as they are
superfluous: the 0.11 versions load fine in 0.12.